refactor: add type safety checks for integer conversions
header.GetOriginalSize now returns int64 and panics on uint64 overflow.
stream.ChunkWriter adds a panic for uint32 chunk size overflow.
utils.ToBytes and FromBytes now use generics, include length checks,
and utilize a configurable byte order.

+++ processor/processor.go
@@ -121,8 +121,7 @@ func (p *Processor) Decrypt(srcPath, destPath, password string) error {
 	}
 
 	originalSize := h.GetOriginalSize()
-	// #nosec G115
-	if err := processor.Process(context.Background(), srcFile, destFile, int64(originalSize)); err != nil {
+	if err := processor.Process(context.Background(), srcFile, destFile, originalSize); err != nil {
 		return fmt.Errorf("failed to process file: %w", err)
 	}
